Added initial Unity Project to test Git
This is the initial commit state of the SWFFRPG_Manager.  The app is
currently functional with few errors.

Known Issues:
- As the Assassin and Outlaw Tech specializations are the only ones with
ChildNodes set, they are the only specializations with working talent
trees.
- Talent trees do not properly lock off "deep" nodes, at the moment.
The idea was, I would make sure the user could not remove talents in the
middle of the hierarchy.  Only leaves.  The proper solution might just
be to not save the "obtainment" of the talent until the whole initial
build is committed to, and then lock off already-committed talents.

TODO:
- Add a specialization purchase option in the Initial Build screen, and
create a secondary specialization button list for switching between
talent trees.
- Add a final build commit/save screen.
- Add save/loading functionality.  Loading will immediately go to the
build screen.